Implement the pretext experiment
