+++ src/lfxai/explanations/features.py
@@ -104,35 +104,6 @@ def proto_attribute(
     for inputs, _ in data_loader:
         inputs = pert(inputs).to(device)
 
-        # Forward the image variable through the network
-        _, distances = ppnet.push_forward(inputs)
-        activation_pattern = ppnet.distance_2_similarity(distances)
-        activation_pattern = activation_pattern.detach().numpy()
+        attributions.append(ppnet.get_feature_importance(inputs, top_k_weights))
 
-        _, min_distances = ppnet(inputs)
-        prototype_activations = (
-            F.softmax(ppnet.distance_2_similarity(min_distances), dim=1)
-            .detach()
-            .numpy()
-        )
-
-        for img, weights in zip(activation_pattern, prototype_activations):
-            # sort and take weighted average of top K prototypes
-            sorted_weights = np.argsort(weights)
-            filter = (weights >= weights[sorted_weights[-top_k_weights]]).astype(float)
-            filtered_weights = filter * weights
-            feature_importance = np.einsum("i,ikl->kl", filtered_weights, img)
-            feature_importance_upsampled = cv2.resize(
-                feature_importance,
-                dsize=(28, 28),
-                interpolation=cv2.INTER_CUBIC,
-            )
-
-            """feature_importance_upsampled_scaled = (
-                feature_importance_upsampled - feature_importance_upsampled.min()
-            ) / (
-                feature_importance_upsampled.max() - feature_importance_upsampled.min()
-            )"""
-            attributions.append(feature_importance_upsampled.reshape(1, 1, 28, 28))
-
-    return np.concatenate(attributions)
+    return np.vstack(attributions)
